AWS Lambda で動かせるようにアプリ構造を変更。
