+++ .chalice/config.json
@@ -1,0 +1,9 @@
+{
+  "version": "2.0",
+  "app_name": "fgo_time_notice",
+  "stages": {
+    "dev": {
+      "api_gateway_stage": "api"
+    }
+  }
+}
