+++ .gitignore
@@ -1,0 +1,3 @@
+.chalice/deployments/
+.chalice/venv/
+.chalice/deployed/
