+++ Makefile
@@ -1,0 +1,7 @@
+.PHONY: deploy lock
+
+deploy:
+	chalice deploy --no-autogen-policy
+
+lock:
+	pipenv lock -r > requirements.txt
