+++ Pipfile
@@ -9,6 +9,7 @@ verify_ssl = true
 beautifulsoup4 = "*"
 requests = "*"
 chalice = "*"
+boto3 = "*"
 
 [requires]
 python_version = "3.8"
