+++ chalicelib/settings.py
@@ -1,0 +1,2 @@
+BUCKET_NAME = 'fgojunks.max747.org'
+JSON_PATH = 'timer/assets/events.json'
