+++ requirements.txt
@@ -1,6 +1,7 @@
 -i https://pypi.org/simple
 attrs==20.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
 beautifulsoup4==4.9.3
+boto3==1.16.11
 botocore==1.19.11
 certifi==2020.6.20
 chalice==1.21.4
@@ -13,6 +14,7 @@ mypy-extensions==0.4.3
 python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
 pyyaml==5.3.1
 requests==2.24.0
+s3transfer==0.3.3
 six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
 soupsieve==2.0.1; python_version >= '3.0'
 urllib3==1.25.11; python_version != '3.4'
